Isolate pruning validation tests
